Move tests into separate module
